Move non-types stuff out of types.ts
- Moves object related methods to `objects.ts`
- Moves assertion to `asserts.ts`
- Move URI stuff to `uri.ts`
