+++ src/vs/platform/files/browser/indexedDBFileSystemProvider.ts
@@ -8,8 +8,8 @@ import { VSBuffer } from 'vs/base/common/buffer';
 import { Emitter, Event } from 'vs/base/common/event';
 import { Disposable, IDisposable } from 'vs/base/common/lifecycle';
 import { ExtUri } from 'vs/base/common/resources';
-import { isString, UriDto } from 'vs/base/common/types';
-import { URI } from 'vs/base/common/uri';
+import { isString } from 'vs/base/common/types';
+import { URI, UriDto } from 'vs/base/common/uri';
 import { localize } from 'vs/nls';
 import { createFileSystemProviderError, FileChangeType, IFileDeleteOptions, IFileOverwriteOptions, FileSystemProviderCapabilities, FileSystemProviderError, FileSystemProviderErrorCode, FileType, IFileWriteOptions, IFileChange, IFileSystemProviderWithFileReadWriteCapability, IStat, IWatchOptions } from 'vs/platform/files/common/files';
 import { DBClosedError, IndexedDB } from 'vs/base/browser/indexedDB';
