+++ src/vs/platform/sharedProcess/node/sharedProcess.ts
@@ -9,7 +9,7 @@ import { NativeParsedArgs } from 'vs/platform/environment/common/argv';
 import { LogLevel } from 'vs/platform/log/common/log';
 import { IUserDataProfile } from 'vs/platform/userDataProfile/common/userDataProfile';
 import { PolicyDefinition, PolicyValue } from 'vs/platform/policy/common/policy';
-import { UriDto } from 'vs/base/common/types';
+import { UriDto } from 'vs/base/common/uri';
 
 export interface ISharedProcess {
