+++ src/vs/platform/userDataProfile/common/userDataProfile.ts
@@ -7,8 +7,8 @@ import { hash } from 'vs/base/common/hash';
 import { Emitter, Event } from 'vs/base/common/event';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { joinPath } from 'vs/base/common/resources';
-import { isUndefined, UriDto } from 'vs/base/common/types';
-import { URI } from 'vs/base/common/uri';
+import { isUndefined } from 'vs/base/common/types';
+import { URI, UriDto } from 'vs/base/common/uri';
 import { localize } from 'vs/nls';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 import { IFileService } from 'vs/platform/files/common/files';
