+++ src/vs/platform/windows/test/electron-main/windowsFinder.test.ts
@@ -8,8 +8,7 @@ import { CancellationToken } from 'vs/base/common/cancellation';
 import { Event } from 'vs/base/common/event';
 import { join } from 'vs/base/common/path';
 import { extUriBiasedIgnorePathCase } from 'vs/base/common/resources';
-import { UriDto } from 'vs/base/common/types';
-import { URI } from 'vs/base/common/uri';
+import { URI, UriDto } from 'vs/base/common/uri';
 import { getPathFromAmdModule } from 'vs/base/test/node/testUtils';
 import { ICommandAction } from 'vs/platform/action/common/action';
 import { NativeParsedArgs } from 'vs/platform/environment/common/argv';
